Move projectName field to Job class
